Data Pipeline Refactor: Move Image Object Detection to new DataModule (#1036)
