+++ flash/core/data/new_data_module.py
@@ -302,10 +302,10 @@ class DataModule(DataModule):
             self._val_input,
             self._test_input,
             self._predict_input,
-            self._train_input.transform,
-            self._val_input.transform,
-            self._test_input.transform,
-            self._predict_input.transform,
+            getattr(self._train_input, "transform", None),
+            getattr(self._val_input, "transform", None),
+            getattr(self._test_input, "transform", None),
+            getattr(self._predict_input, "transform", None),
             task._deserializer,
             task._output_transform,
             task._output,
