+++ flash/image/detection/data.py
@@ -11,18 +11,17 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from functools import partial
-from typing import Any, Callable, Dict, Hashable, List, Optional, Sequence, Tuple, Type, TYPE_CHECKING, Union
+from typing import Any, Dict, List, Optional, Type, TYPE_CHECKING
 
-from flash.core.data.data_module import DataModule
 from flash.core.data.data_pipeline import DataPipelineState
-from flash.core.data.io.input import DataKeys, InputFormat
-from flash.core.data.io.input_transform import InputTransform
-from flash.core.integrations.fiftyone.utils import FiftyOneLabelUtilities
+from flash.core.data.io.input_base import Input
+from flash.core.data.new_data_module import DataModule
 from flash.core.integrations.icevision.data import IceVisionInput
-from flash.core.integrations.icevision.transforms import default_transforms
+from flash.core.integrations.icevision.transforms import IceVisionTransform
 from flash.core.utilities.imports import _FIFTYONE_AVAILABLE, _ICEVISION_AVAILABLE, lazy_import, requires
 from flash.core.utilities.stages import RunningStage
+from flash.core.utilities.types import INPUT_TRANSFORM_TYPE
+from flash.image.detection.input import ObjectDetectionFiftyOneInput
 
 SampleCollection = None
 if _FIFTYONE_AVAILABLE:
@@ -33,10 +32,7 @@ else:
     foc, fol = None, None
 
 if _ICEVISION_AVAILABLE:
-    from icevision.core import BBox, ClassMap, IsCrowdsRecordComponent, ObjectDetectionRecord
-    from icevision.data import SingleSplitSplitter
     from icevision.parsers import COCOBBoxParser, Parser, VIABBoxParser, VOCBBoxParser
-    from icevision.utils import ImgSize
 else:
     COCOBBoxParser = object
     VIABBoxParser = object
@@ -44,141 +40,9 @@ else:
     Parser = object
 
 
-class FiftyOneParser(Parser):
-    def __init__(self, data, class_map, label_field, iscrowd):
-        template_record = ObjectDetectionRecord()
-        template_record.add_component(IsCrowdsRecordComponent())
-        super().__init__(template_record=template_record)
-
-        data = data
-        label_field = label_field
-        iscrowd = iscrowd
-
-        self.data = []
-        self.class_map = class_map
-
-        for fp, w, h, sample_labs, sample_boxes, sample_iscrowd in zip(
-            data.values("filepath"),
-            data.values("metadata.width"),
-            data.values("metadata.height"),
-            data.values(label_field + ".detections.label"),
-            data.values(label_field + ".detections.bounding_box"),
-            data.values(label_field + ".detections." + iscrowd),
-        ):
-            for lab, box, iscrowd in zip(sample_labs, sample_boxes, sample_iscrowd):
-                self.data.append((fp, w, h, lab, box, iscrowd))
-
-    def __iter__(self) -> Any:
-        return iter(self.data)
-
-    def __len__(self) -> int:
-        return len(self.data)
-
-    def record_id(self, o) -> Hashable:
-        return o[0]
-
-    def parse_fields(self, o, record, is_new):
-        fp, w, h, lab, box, iscrowd = o
-
-        if iscrowd is None:
-            iscrowd = 0
-
-        if is_new:
-            record.set_filepath(fp)
-            record.set_img_size(ImgSize(width=w, height=h))
-            record.detection.set_class_map(self.class_map)
-
-        box = self._reformat_bbox(*box, w, h)
-
-        record.detection.add_bboxes([BBox.from_xyxy(*box)])
-        record.detection.add_labels([lab])
-        record.detection.add_iscrowds([iscrowd])
-
-    @staticmethod
-    def _reformat_bbox(xmin, ymin, box_w, box_h, img_w, img_h):
-        xmin *= img_w
-        ymin *= img_h
-        box_w *= img_w
-        box_h *= img_h
-        xmax = xmin + box_w
-        ymax = ymin + box_h
-        output_bbox = [xmin, ymin, xmax, ymax]
-        return output_bbox
-
-
-class ObjectDetectionFiftyOneInput(IceVisionInput):
-    @requires("fiftyone")
-    def load_data(
-        self,
-        sample_collection: SampleCollection,
-        label_field: str = "ground_truth",
-        iscrowd: str = "iscrowd",
-    ) -> Sequence[Dict[str, Any]]:
-        label_utilities = FiftyOneLabelUtilities(label_field, fol.Detections)
-        label_utilities.validate(sample_collection)
-        sample_collection.compute_metadata()
-        classes = label_utilities.get_classes(sample_collection)
-        class_map = ClassMap(classes)
-        self.num_classes = len(class_map)
-
-        parser = FiftyOneParser(sample_collection, class_map, label_field, iscrowd)
-        records = parser.parse(data_splitter=SingleSplitSplitter())
-        return [{DataKeys.INPUT: record} for record in records[0]]
-
-    @staticmethod
-    @requires("fiftyone")
-    def predict_load_data(sample_collection: SampleCollection) -> Sequence[Dict[str, Any]]:
-        return [{DataKeys.INPUT: f} for f in sample_collection.values("filepath")]
-
-
-class ObjectDetectionInputTransform(InputTransform):
-    def __init__(
-        self,
-        train_transform: Optional[Dict[str, Callable]] = None,
-        val_transform: Optional[Dict[str, Callable]] = None,
-        test_transform: Optional[Dict[str, Callable]] = None,
-        predict_transform: Optional[Dict[str, Callable]] = None,
-        image_size: Tuple[int, int] = (128, 128),
-        parser: Optional[Callable] = None,
-    ):
-        self.image_size = image_size
-
-        super().__init__(
-            train_transform=train_transform,
-            val_transform=val_transform,
-            test_transform=test_transform,
-            predict_transform=predict_transform,
-            inputs={
-                "coco": partial(IceVisionInput, parser=COCOBBoxParser),
-                "via": partial(IceVisionInput, parser=VIABBoxParser),
-                "voc": partial(IceVisionInput, parser=VOCBBoxParser),
-                "icedata": partial(IceVisionInput, parser=parser),
-                InputFormat.FILES: IceVisionInput,
-                InputFormat.FOLDERS: partial(IceVisionInput, parser=parser),
-                InputFormat.FIFTYONE: ObjectDetectionFiftyOneInput,
-            },
-            default_input="icedata",
-        )
-
-        self._default_collate = self._identity
-
-    def get_state_dict(self) -> Dict[str, Any]:
-        return {**self.transforms}
-
-    @classmethod
-    def load_state_dict(cls, state_dict: Dict[str, Any], strict: bool = False):
-        return cls(**state_dict)
-
-    def default_transforms(self) -> Optional[Dict[str, Callable]]:
-        return default_transforms(self.image_size)
-
-    def train_default_transforms(self) -> Optional[Dict[str, Callable]]:
-        return default_transforms(self.image_size)
-
-
 class ObjectDetectionData(DataModule):
 
-    input_transform_cls = ObjectDetectionInputTransform
+    input_transform_cls = IceVisionTransform
 
     @classmethod
     def from_icedata(
@@ -190,29 +54,23 @@ class ObjectDetectionData(DataModule):
         test_folder: Optional[str] = None,
         test_ann_file: Optional[str] = None,
         predict_folder: Optional[str] = None,
-        train_transform: Optional[Dict[str, Callable]] = None,
-        val_transform: Optional[Dict[str, Callable]] = None,
-        test_transform: Optional[Dict[str, Callable]] = None,
-        predict_transform: Optional[Dict[str, Callable]] = None,
-        image_size: Tuple[int, int] = (128, 128),
-        parser: Optional[Union[Callable, Type[Parser]]] = None,
+        train_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
+        val_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
+        test_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
+        predict_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
+        parser: Type[Parser] = Parser,
+        input_cls: Type[Input] = IceVisionInput,
+        transform_kwargs: Optional[Dict] = None,
         **data_module_kwargs,
     ) -> "ObjectDetectionData":
 
-        dataset_kwargs = dict(parser=parser, data_pipeline_state=DataPipelineState())
+        ds_kw = dict(parser=parser, data_pipeline_state=DataPipelineState(), transform_kwargs=transform_kwargs)
 
         return cls(
-            IceVisionInput(RunningStage.TRAINING, train_folder, train_ann_file, **dataset_kwargs),
-            IceVisionInput(RunningStage.VALIDATING, val_folder, val_ann_file, **dataset_kwargs),
-            IceVisionInput(RunningStage.TESTING, test_folder, test_ann_file, **dataset_kwargs),
-            IceVisionInput(RunningStage.PREDICTING, predict_folder, **dataset_kwargs),
-            input_transform=cls.input_transform_cls(
-                train_transform,
-                val_transform,
-                test_transform,
-                predict_transform,
-                image_size=image_size,
-            ),
+            input_cls(RunningStage.TRAINING, train_folder, train_ann_file, transform=train_transform, **ds_kw),
+            input_cls(RunningStage.VALIDATING, val_folder, val_ann_file, transform=val_transform, **ds_kw),
+            input_cls(RunningStage.TESTING, test_folder, test_ann_file, transform=test_transform, **ds_kw),
+            input_cls(RunningStage.PREDICTING, predict_folder, transform=predict_transform, **ds_kw),
             **data_module_kwargs,
         )
 
@@ -226,11 +84,12 @@ class ObjectDetectionData(DataModule):
         test_folder: Optional[str] = None,
         test_ann_file: Optional[str] = None,
         predict_folder: Optional[str] = None,
-        train_transform: Optional[Dict[str, Callable]] = None,
-        val_transform: Optional[Dict[str, Callable]] = None,
-        test_transform: Optional[Dict[str, Callable]] = None,
-        predict_transform: Optional[Dict[str, Callable]] = None,
-        image_size: Tuple[int, int] = (128, 128),
+        train_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
+        val_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
+        test_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
+        predict_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
+        input_cls: Type[Input] = IceVisionInput,
+        transform_kwargs: Optional[Dict] = None,
         **data_module_kwargs: Any,
     ) -> "ObjectDetectionData":
         """Creates a :class:`~flash.image.detection.data.ObjectDetectionData` object from the given data folders
@@ -252,7 +111,9 @@ class ObjectDetectionData(DataModule):
                 :class:`~flash.core.data.io.input_transform.InputTransform` hook names to callable transforms.
             predict_transform: The dictionary of transforms to use during predicting which maps
                 :class:`~flash.core.data.io.input_transform.InputTransform` hook names to callable transforms.
-            image_size: The size to resize images (and their bounding boxes) to.
+            input_cls: The :class:`~flash.core.data.io.input_base.Input` used to create the dataset.
+            transform_kwargs: Keyword arguments provided to the transform on instantiation.
+            data_module_kwargs: Keyword arguments provided to the DataModule on instantiation.
         """
         return cls.from_icedata(
             train_folder=train_folder,
@@ -266,8 +127,9 @@ class ObjectDetectionData(DataModule):
             val_transform=val_transform,
             test_transform=test_transform,
             predict_transform=predict_transform,
-            image_size=image_size,
             parser=COCOBBoxParser,
+            input_cls=input_cls,
+            transform_kwargs=transform_kwargs,
             **data_module_kwargs,
         )
 
@@ -281,11 +143,12 @@ class ObjectDetectionData(DataModule):
         test_folder: Optional[str] = None,
         test_ann_file: Optional[str] = None,
         predict_folder: Optional[str] = None,
-        train_transform: Optional[Dict[str, Callable]] = None,
-        val_transform: Optional[Dict[str, Callable]] = None,
-        test_transform: Optional[Dict[str, Callable]] = None,
-        predict_transform: Optional[Dict[str, Callable]] = None,
-        image_size: Tuple[int, int] = (128, 128),
+        train_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
+        val_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
+        test_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
+        predict_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
+        input_cls: Type[Input] = IceVisionInput,
+        transform_kwargs: Optional[Dict] = None,
         **data_module_kwargs: Any,
     ) -> "ObjectDetectionData":
         """Creates a :class:`~flash.image.detection.data.ObjectDetectionData` object from the given data folders
@@ -307,7 +170,9 @@ class ObjectDetectionData(DataModule):
                 :class:`~flash.core.data.io.input_transform.InputTransform` hook names to callable transforms.
             predict_transform: The dictionary of transforms to use during predicting which maps
                 :class:`~flash.core.data.io.input_transform.InputTransform` hook names to callable transforms.
-            image_size: The size to resize images (and their bounding boxes) to.
+            input_cls: The :class:`~flash.core.data.io.input_base.Input` used to create the dataset.
+            transform_kwargs: Keyword arguments provided to the transform on instantiation.
+            data_module_kwargs: Keyword arguments provided to the DataModule on instantiation.
         """
         return cls.from_icedata(
             train_folder=train_folder,
@@ -321,8 +186,9 @@ class ObjectDetectionData(DataModule):
             val_transform=val_transform,
             test_transform=test_transform,
             predict_transform=predict_transform,
-            image_size=image_size,
             parser=VOCBBoxParser,
+            input_cls=input_cls,
+            transform_kwargs=transform_kwargs,
             **data_module_kwargs,
         )
 
@@ -336,11 +202,12 @@ class ObjectDetectionData(DataModule):
         test_folder: Optional[str] = None,
         test_ann_file: Optional[str] = None,
         predict_folder: Optional[str] = None,
-        train_transform: Optional[Dict[str, Callable]] = None,
-        val_transform: Optional[Dict[str, Callable]] = None,
-        test_transform: Optional[Dict[str, Callable]] = None,
-        predict_transform: Optional[Dict[str, Callable]] = None,
-        image_size: Tuple[int, int] = (128, 128),
+        train_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
+        val_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
+        test_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
+        predict_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
+        input_cls: Type[Input] = IceVisionInput,
+        transform_kwargs: Optional[Dict] = None,
         **data_module_kwargs: Any,
     ) -> "ObjectDetectionData":
         """Creates a :class:`~flash.image.detection.data.ObjectDetectionData` object from the given data folders
@@ -362,7 +229,9 @@ class ObjectDetectionData(DataModule):
                 :class:`~flash.core.data.io.input_transform.InputTransform` hook names to callable transforms.
             predict_transform: The dictionary of transforms to use during predicting which maps
                 :class:`~flash.core.data.io.input_transform.InputTransform` hook names to callable transforms.
-            image_size: The size to resize images (and their bounding boxes) to.
+            input_cls: The :class:`~flash.core.data.io.input_base.Input` used to create the dataset.
+            transform_kwargs: Keyword arguments provided to the transform on instantiation.
+            data_module_kwargs: Keyword arguments provided to the DataModule on instantiation.
         """
         return cls.from_icedata(
             train_folder=train_folder,
@@ -376,8 +245,9 @@ class ObjectDetectionData(DataModule):
             val_transform=val_transform,
             test_transform=test_transform,
             predict_transform=predict_transform,
-            image_size=image_size,
             parser=VIABBoxParser,
+            input_cls=input_cls,
+            transform_kwargs=transform_kwargs,
             **data_module_kwargs,
         )
 
@@ -389,26 +259,25 @@ class ObjectDetectionData(DataModule):
         val_dataset: Optional[SampleCollection] = None,
         test_dataset: Optional[SampleCollection] = None,
         predict_dataset: Optional[SampleCollection] = None,
-        train_transform: Optional[Dict[str, Callable]] = None,
-        val_transform: Optional[Dict[str, Callable]] = None,
-        test_transform: Optional[Dict[str, Callable]] = None,
-        predict_transform: Optional[Dict[str, Callable]] = None,
+        train_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
+        val_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
+        test_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
+        predict_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
         label_field: str = "ground_truth",
         iscrowd: str = "iscrowd",
-        image_size: Tuple[int, int] = (128, 128),
-        **data_module_kwargs,
+        input_cls: Type[Input] = ObjectDetectionFiftyOneInput,
+        transform_kwargs: Optional[Dict] = None,
+        **data_module_kwargs: Any,
     ) -> "ObjectDetectionData":
+
+        ds_kw = dict(data_pipeline_state=DataPipelineState(), transform_kwargs=transform_kwargs)
+
         return cls(
-            ObjectDetectionFiftyOneInput(RunningStage.TRAINING, train_dataset, label_field, iscrowd),
-            ObjectDetectionFiftyOneInput(RunningStage.VALIDATING, val_dataset, label_field, iscrowd),
-            ObjectDetectionFiftyOneInput(RunningStage.TESTING, test_dataset, label_field, iscrowd),
-            ObjectDetectionFiftyOneInput(RunningStage.PREDICTING, predict_dataset, label_field, iscrowd),
-            input_transform=cls.input_transform_cls(
-                train_transform,
-                val_transform,
-                test_transform,
-                predict_transform,
-                image_size=image_size,
+            input_cls(RunningStage.TRAINING, train_dataset, label_field, iscrowd, transform=train_transform, **ds_kw),
+            input_cls(RunningStage.VALIDATING, val_dataset, label_field, iscrowd, transform=val_transform, **ds_kw),
+            input_cls(RunningStage.TESTING, test_dataset, label_field, iscrowd, transform=test_transform, **ds_kw),
+            input_cls(
+                RunningStage.PREDICTING, predict_dataset, label_field, iscrowd, transform=predict_transform, **ds_kw
             ),
             **data_module_kwargs,
         )
@@ -417,8 +286,9 @@ class ObjectDetectionData(DataModule):
     def from_folders(
         cls,
         predict_folder: Optional[str] = None,
-        predict_transform: Optional[Dict[str, Callable]] = None,
-        image_size: Tuple[int, int] = (128, 128),
+        predict_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
+        input_cls: Type[Input] = IceVisionInput,
+        transform_kwargs: Optional[Dict] = None,
         **data_module_kwargs: Any,
     ) -> "DataModule":
         """Creates a :class:`~flash.image.detection.data.ObjectDetectionData` object from the given data folders
@@ -433,10 +303,8 @@ class ObjectDetectionData(DataModule):
             The constructed data module.
         """
         return cls(
-            predict_dataset=IceVisionInput(RunningStage.PREDICTING, predict_folder),
-            input_transform=cls.input_transform_cls(
-                predict_transform=predict_transform,
-                image_size=image_size,
+            predict_input=input_cls(
+                RunningStage.PREDICTING, predict_folder, transform=predict_transform, transform_kwargs=transform_kwargs
             ),
             **data_module_kwargs,
         )
@@ -445,8 +313,9 @@ class ObjectDetectionData(DataModule):
     def from_files(
         cls,
         predict_files: Optional[List[str]] = None,
-        predict_transform: Optional[Dict[str, Callable]] = None,
-        image_size: Tuple[int, int] = (128, 128),
+        predict_transform: INPUT_TRANSFORM_TYPE = IceVisionTransform,
+        input_cls: Type[Input] = IceVisionInput,
+        transform_kwargs: Optional[Dict] = None,
         **data_module_kwargs: Any,
     ) -> "DataModule":
         """Creates a :class:`~flash.image.detection.data.ObjectDetectionData` object from the given data files.
@@ -462,10 +331,8 @@ class ObjectDetectionData(DataModule):
             The constructed data module.
         """
         return cls(
-            predict_dataset=IceVisionInput(RunningStage.PREDICTING, predict_files),
-            input_transform=cls.input_transform_cls(
-                predict_transform=predict_transform,
-                image_size=image_size,
+            predict_input=input_cls(
+                RunningStage.PREDICTING, predict_files, transform=predict_transform, transform_kwargs=transform_kwargs
             ),
             **data_module_kwargs,
         )
